apps/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.9
python-jose[cryptography]==3.5.0
passlib==1.7.4
python-multipart==0.0.20
httpx==0.28.1
openai==1.68.2
